+++ src/Files.App/Helpers/Environment/ElevationHelpers.cs
@@ -8,12 +8,15 @@ namespace Files.App.Helpers
 {
 	public static class ElevationHelpers
 	{
+		[DllImport("shell32.dll", EntryPoint = "#865", CharSet = CharSet.Unicode, SetLastError = true)]
+		[return: MarshalAs(UnmanagedType.Bool)] private static extern bool _IsElevationRequired([MarshalAs(UnmanagedType.LPWStr)] string pszPath);
+
 		public static bool IsElevationRequired(string filePath)
 		{
 			if (string.IsNullOrEmpty(filePath))
 				return false;
 
-			return Win32PInvoke._IsElevationRequired(filePath);
+			return _IsElevationRequired(filePath);
 		}
 
 		public static bool IsAppRunAsAdmin()
