+++ src/Files.App/Helpers/Win32/Win32PInvoke.Methods.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) 2024 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using System.IO;
 using System.Runtime.InteropServices;
-using System.Runtime.InteropServices.ComTypes;
 using System.Text;
 using Vanara.PInvoke;
 using static Vanara.PInvoke.User32;
@@ -17,11 +15,6 @@ namespace Files.App.Helpers
 			uint dwNumberOfBytesTransfered,
 			OVERLAPPED lpOverlapped
 		);
-
-		public delegate void LPOVERLAPPED_COMPLETION_ROUTINE(
-			uint dwErrorCode,
-			uint dwNumberOfBytesTransfered,
-			ref NativeOverlapped lpOverlapped);
 
 		[DllImport("rstrtmgr.dll", CharSet = CharSet.Unicode)]
 		public static extern int RmRegisterResources(
@@ -189,7 +182,7 @@ namespace Files.App.Helpers
 			uint dwIoControlCode,
 			IntPtr lpInBuffer,
 			uint nInBufferSize,
-			out IntPtr lpOutBuffer,
+			IntPtr lpOutBuffer,
 			uint nOutBufferSize,
 			out uint lpBytesReturned,
 			IntPtr lpOverlapped
@@ -216,194 +209,5 @@ namespace Files.App.Helpers
 			int bufferSize,
 			uint flags
 		);
-
-		[DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
-		CallingConvention = CallingConvention.StdCall,
-		SetLastError = true)]
-		public static extern IntPtr CreateFileFromApp(
-			string lpFileName,
-			uint dwDesiredAccess,
-			uint dwShareMode,
-			IntPtr SecurityAttributes,
-			uint dwCreationDisposition,
-			uint dwFlagsAndAttributes,
-			IntPtr hTemplateFile
-		);
-
-		[DllImport("api-ms-win-core-io-l1-1-0.dll", ExactSpelling = true, SetLastError = true, CharSet = CharSet.Auto)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		private static extern bool DeviceIoControl(
-			IntPtr hDevice,
-			uint dwIoControlCode,
-			IntPtr lpInBuffer,
-			uint nInBufferSize,
-			//IntPtr lpOutBuffer,
-			out REPARSE_DATA_BUFFER outBuffer,
-			uint nOutBufferSize,
-			out uint lpBytesReturned,
-			IntPtr lpOverlapped);
-
-		[DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", SetLastError = true, CharSet = CharSet.Auto)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool GetFileAttributesExFromApp(
-			string lpFileName,
-			GET_FILEEX_INFO_LEVELS fInfoLevelId,
-			out WIN32_FILE_ATTRIBUTE_DATA lpFileInformation);
-
-		[DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", SetLastError = true, CharSet = CharSet.Auto)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool SetFileAttributesFromApp(
-			string lpFileName,
-			FileAttributes dwFileAttributes);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", ExactSpelling = true,
-		CallingConvention = CallingConvention.StdCall,
-		SetLastError = true)]
-		public static extern uint SetFilePointer(
-			IntPtr hFile,
-			long lDistanceToMove,
-			IntPtr lpDistanceToMoveHigh,
-			uint dwMoveMethod
-		);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", CharSet = CharSet.Auto,
-		CallingConvention = CallingConvention.StdCall,
-		SetLastError = true)]
-		public unsafe static extern bool ReadFile(
-			IntPtr hFile,
-			byte* lpBuffer,
-			int nBufferLength,
-			int* lpBytesReturned,
-			IntPtr lpOverlapped
-		);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", CharSet = CharSet.Auto,
-		CallingConvention = CallingConvention.StdCall,
-		SetLastError = true)]
-		public unsafe static extern bool WriteFile(
-			IntPtr hFile,
-			byte* lpBuffer,
-			int nBufferLength,
-			int* lpBytesWritten,
-			IntPtr lpOverlapped
-		);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", CharSet = CharSet.Auto,
-			CallingConvention = CallingConvention.StdCall,
-			SetLastError = true)]
-		public static extern bool WriteFileEx(
-			IntPtr hFile,
-			byte[] lpBuffer,
-			uint nNumberOfBytesToWrite,
-			[In] ref NativeOverlapped lpOverlapped,
-			LPOVERLAPPED_COMPLETION_ROUTINE lpCompletionRoutine);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
-		public static extern bool GetFileTime([In] IntPtr hFile, out FILETIME lpCreationTime, out FILETIME lpLastAccessTime, out FILETIME lpLastWriteTime);
-
-		[DllImport("api-ms-win-core-file-l1-2-1.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
-		public static extern bool SetFileTime([In] IntPtr hFile, in FILETIME lpCreationTime, in FILETIME lpLastAccessTime, in FILETIME lpLastWriteTime);
-
-		[DllImport("api-ms-win-core-file-l2-1-1.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
-		public static extern bool GetFileInformationByHandleEx(IntPtr hFile, FILE_INFO_BY_HANDLE_CLASS infoClass, out FILE_ID_BOTH_DIR_INFO dirInfo, uint dwBufferSize);
-
-		[DllImport("api-ms-win-core-file-l2-1-1.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
-		public static extern bool GetFileInformationByHandleEx(IntPtr hFile, FILE_INFO_BY_HANDLE_CLASS infoClass, IntPtr dirInfo, uint dwBufferSize);
-
-		[DllImport("kernel32.dll", ExactSpelling = true, CharSet = CharSet.Auto, SetLastError = true)]
-		public static extern IntPtr FindFirstStreamW(string lpFileName, StreamInfoLevels InfoLevel, [In, Out, MarshalAs(UnmanagedType.LPStruct)] WIN32_FIND_STREAM_DATA lpFindStreamData, uint dwFlags);
-
-		[DllImport("kernel32.dll", ExactSpelling = true, CharSet = CharSet.Auto, SetLastError = true)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool FindNextStreamW(IntPtr hndFindFile, [In, Out, MarshalAs(UnmanagedType.LPStruct)] WIN32_FIND_STREAM_DATA lpFindStreamData);
-
-		[DllImport("Shcore.dll", SetLastError = true)]
-		public static extern int GetDpiForMonitor(IntPtr hmonitor, uint dpiType, out uint dpiX, out uint dpiY);
-
-		[DllImport("api-ms-win-core-processthreads-l1-1-0.dll", SetLastError = true, ExactSpelling = true)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool OpenProcessToken([In] IntPtr ProcessHandle, TokenAccess DesiredAccess, out IntPtr TokenHandle);
-
-		[DllImport("api-ms-win-core-processthreads-l1-1-2.dll", SetLastError = true, ExactSpelling = true)]
-		public static extern IntPtr GetCurrentProcess();
-
-		[DllImport("api-ms-win-security-base-l1-1-0.dll", SetLastError = true, ExactSpelling = true)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool GetTokenInformation(IntPtr hObject, TOKEN_INFORMATION_CLASS tokenInfoClass, IntPtr pTokenInfo, int tokenInfoLength, out int returnLength);
-
-		[DllImport("api-ms-win-security-base-l1-1-0.dll", ExactSpelling = true, SetLastError = true)]
-		public static extern int GetLengthSid(IntPtr pSid);
-
-		[DllImport("crypt32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool CryptUnprotectData(
-			in CRYPTOAPI_BLOB pDataIn,
-			StringBuilder szDataDescr,
-			in CRYPTOAPI_BLOB pOptionalEntropy,
-			IntPtr pvReserved,
-			IntPtr pPromptStruct,
-			CryptProtectFlags dwFlags,
-			out CRYPTOAPI_BLOB pDataOut);
-
-		[DllImport("api-ms-win-core-wow64-l1-1-1.dll", SetLastError = true)]
-		public static extern bool IsWow64Process2(
-			IntPtr process,
-			out ushort processMachine,
-			out ushort nativeMachine);
-
-		[DllImport("api-ms-win-core-file-l1-1-0.dll", CharSet = CharSet.Unicode)]
-		public static extern bool FindNextFile(
-			IntPtr hFindFile,
-			out WIN32_FIND_DATA lpFindFileData);
-
-		[DllImport("api-ms-win-core-file-l1-1-0.dll")]
-		public static extern bool FindClose(
-			IntPtr hFindFile);
-
-		[DllImport("api-ms-win-core-timezone-l1-1-0.dll", SetLastError = true)]
-		public static extern bool FileTimeToSystemTime(
-			ref FILETIME lpFileTime,
-			out SYSTEMTIME lpSystemTime);
-
-		[DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", SetLastError = true, CharSet = CharSet.Unicode)]
-		public static extern IntPtr FindFirstFileExFromApp(
-			string lpFileName,
-			FINDEX_INFO_LEVELS fInfoLevelId,
-			out WIN32_FIND_DATA lpFindFileData,
-			FINDEX_SEARCH_OPS fSearchOp,
-			IntPtr lpSearchFilter,
-			int dwAdditionalFlags);
-
-		[DllImport("api-ms-win-core-string-l1-1-0.dll", CharSet = CharSet.Unicode)]
-		public static extern int CompareStringEx(
-			string localeName,
-			int flags,
-			string str1,
-			int count1,
-			string str2,
-			int count2,
-			IntPtr versionInformation,
-			IntPtr reserved,
-			int param
-		);
-
-		[DllImport("shell32.dll", EntryPoint = "#865", CharSet = CharSet.Unicode, SetLastError = true)]
-		[return: MarshalAs(UnmanagedType.Bool)]
-		public static extern bool _IsElevationRequired([MarshalAs(UnmanagedType.LPWStr)] string pszPath);
-
-		[DllImport("shlwapi.dll", CallingConvention = CallingConvention.StdCall, PreserveSig = true, CharSet = CharSet.Unicode)]
-		public static extern HRESULT SHCreateStreamOnFileEx(string pszFile, STGM grfMode, uint dwAttributes, uint fCreate, IntPtr pstmTemplate, out IntPtr ppstm);
-
-		[DllImport("shell32.dll", CallingConvention = CallingConvention.StdCall, PreserveSig = true, CharSet = CharSet.Unicode)]
-		public static extern HRESULT SHCreateItemFromParsingName(string pszPath, IntPtr pbc, ref Guid riid, out IntPtr ppv);
-
-		[DllImport("ole32.dll", CallingConvention = CallingConvention.StdCall)]
-		public static extern HRESULT CoCreateInstance(ref Guid rclsid, IntPtr pUnkOuter, ClassContext dwClsContext, ref Guid riid, out IntPtr ppv);
-
-		[DllImport("kernel32.dll", CharSet = CharSet.Unicode)]
-		public static extern uint RegisterApplicationRestart(string pwzCommandLine, int dwFlags);
-
-		[DllImport(Lib.Shell32, SetLastError = false, CharSet = CharSet.Unicode)]
-		public static extern int SHQueryRecycleBin(string pszRootPath, ref SHQUERYRBINFO pSHQueryRBInfo);
 	}
 }
