+++ src/Files.App/Utils/FileTags/FileTagsHelper.cs
@@ -18,17 +18,17 @@ namespace Files.App.Utils.FileTags
 
 		public static string[] ReadFileTag(string filePath)
 		{
-			var tagString = Win32Helper.ReadStringFromFile($"{filePath}:files");
+			var tagString = NativeFileOperationsHelper.ReadStringFromFile($"{filePath}:files");
 			return tagString?.Split(',', StringSplitOptions.RemoveEmptyEntries) ?? [];
 		}
 
 		public static async void WriteFileTag(string filePath, string[] tag)
 		{
-			var isDateOk = Win32Helper.GetFileDateModified(filePath, out var dateModified); // Backup date modified
-			var isReadOnly = Win32Helper.HasFileAttribute(filePath, IO.FileAttributes.ReadOnly);
+			var isDateOk = NativeFileOperationsHelper.GetFileDateModified(filePath, out var dateModified); // Backup date modified
+			var isReadOnly = NativeFileOperationsHelper.HasFileAttribute(filePath, IO.FileAttributes.ReadOnly);
 			if (isReadOnly) // Unset read-only attribute (#7534)
 			{
-				Win32Helper.UnsetFileAttribute(filePath, IO.FileAttributes.ReadOnly);
+				NativeFileOperationsHelper.UnsetFileAttribute(filePath, IO.FileAttributes.ReadOnly);
 			}
 			if (!tag.Any())
 			{
@@ -36,7 +36,7 @@ namespace Files.App.Utils.FileTags
 			}
 			else if (ReadFileTag(filePath) is not string[] arr || !tag.SequenceEqual(arr))
 			{
-				var result = Win32Helper.WriteStringToFile($"{filePath}:files", string.Join(',', tag));
+				var result = NativeFileOperationsHelper.WriteStringToFile($"{filePath}:files", string.Join(',', tag));
 				if (result == false)
 				{
 					ContentDialog dialog = new()
@@ -54,11 +54,11 @@ namespace Files.App.Utils.FileTags
 			}
 			if (isReadOnly) // Restore read-only attribute (#7534)
 			{
-				Win32Helper.SetFileAttribute(filePath, IO.FileAttributes.ReadOnly);
+				NativeFileOperationsHelper.SetFileAttribute(filePath, IO.FileAttributes.ReadOnly);
 			}
 			if (isDateOk)
 			{
-				Win32Helper.SetFileDateModified(filePath, dateModified); // Restore date modified
+				NativeFileOperationsHelper.SetFileDateModified(filePath, dateModified); // Restore date modified
 			}
 		}
 
@@ -105,7 +105,7 @@ namespace Files.App.Utils.FileTags
 			}
 		}
 
-		public static ulong? GetFileFRN(string filePath) => Win32Helper.GetFileFRN(filePath);
+		public static ulong? GetFileFRN(string filePath) => NativeFileOperationsHelper.GetFileFRN(filePath);
 
 		public static Task<ulong?> GetFileFRN(IStorageItem item)
 		{
