+++ src/Files.App/Utils/Storage/Helpers/StorageHelpers.cs
@@ -32,7 +32,7 @@ namespace Files.App.Helpers
 			}
 
 			// Fast get attributes
-			bool exists = Win32Helper.GetFileAttributesExFromApp(path, Win32Helper.GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out Win32Helper.WIN32_FILE_ATTRIBUTE_DATA itemAttributes);
+			bool exists = NativeFileOperationsHelper.GetFileAttributesExFromApp(path, NativeFileOperationsHelper.GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out NativeFileOperationsHelper.WIN32_FILE_ATTRIBUTE_DATA itemAttributes);
 			if (exists) // Exists on local storage
 			{
 				// Directory
@@ -160,7 +160,7 @@ namespace Files.App.Helpers
 
 		public static bool Exists(string path)
 		{
-			return Win32Helper.GetFileAttributesExFromApp(path, Win32Helper.GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out _);
+			return NativeFileOperationsHelper.GetFileAttributesExFromApp(path, NativeFileOperationsHelper.GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out _);
 		}
 
 		public static IStorageItemWithPath FromStorageItem(this IStorageItem item, string customPath = null, FilesystemItemType? itemType = null)
