+++ src/Files.App/Utils/Storage/Operations/FilesystemOperations.cs
@@ -186,10 +186,10 @@ namespace Files.App.Utils.Storage
 
 						if (fsCopyResult)
 						{
-							if (Win32Helper.HasFileAttribute(source.Path, SystemIO.FileAttributes.Hidden))
+							if (NativeFileOperationsHelper.HasFileAttribute(source.Path, SystemIO.FileAttributes.Hidden))
 							{
 								// The source folder was hidden, apply hidden attribute to destination
-								Win32Helper.SetFileAttribute(fsCopyResult.Result.Path, SystemIO.FileAttributes.Hidden);
+								NativeFileOperationsHelper.SetFileAttribute(fsCopyResult.Result.Path, SystemIO.FileAttributes.Hidden);
 							}
 
 							copiedItem = (BaseStorageFolder)fsCopyResult;
@@ -403,10 +403,10 @@ namespace Files.App.Utils.Storage
 
 							if (fsResultMove)
 							{
-								if (Win32Helper.HasFileAttribute(source.Path, SystemIO.FileAttributes.Hidden))
+								if (NativeFileOperationsHelper.HasFileAttribute(source.Path, SystemIO.FileAttributes.Hidden))
 								{
 									// The source folder was hidden, apply hidden attribute to destination
-									Win32Helper.SetFileAttribute(fsResultMove.Result.Path, SystemIO.FileAttributes.Hidden);
+									NativeFileOperationsHelper.SetFileAttribute(fsResultMove.Result.Path, SystemIO.FileAttributes.Hidden);
 								}
 
 								movedItem = (BaseStorageFolder)fsResultMove;
