+++ src/Files.App/Utils/Storage/StorageItems/ZipStorageFile.cs
@@ -105,7 +105,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenAsync(accessMode);
 					}
 
-					var file = Win32Helper.OpenFileForRead(containerPath, rw);
+					var file = NativeFileOperationsHelper.OpenFileForRead(containerPath, rw);
 					return file.IsInvalid ? null : new FileStream(file, rw ? FileAccess.ReadWrite : FileAccess.Read).AsRandomAccessStream();
 				}
 
@@ -150,7 +150,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenReadAsync();
 					}
 
-					var hFile = Win32Helper.OpenFileForRead(containerPath);
+					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath);
 					return hFile.IsInvalid ? null : new StreamWithContentType(new FileStream(hFile, FileAccess.Read).AsRandomAccessStream());
 				}
 
@@ -189,7 +189,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenSequentialReadAsync();
 					}
 
-					var hFile = Win32Helper.OpenFileForRead(containerPath);
+					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath);
 					return hFile.IsInvalid ? null : new FileStream(hFile, FileAccess.Read).AsInputStream();
 				}
 
@@ -400,7 +400,7 @@ namespace Files.App.Utils.Storage
 		{
 			try
 			{
-				var hFile = Win32Helper.OpenFileForRead(path);
+				var hFile = NativeFileOperationsHelper.OpenFileForRead(path);
 				if (hFile.IsInvalid)
 				{
 					return false;
@@ -475,7 +475,7 @@ namespace Files.App.Utils.Storage
 				}
 				else
 				{
-					var hFile = Win32Helper.OpenFileForRead(containerPath, readWrite);
+					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath, readWrite);
 					if (hFile.IsInvalid)
 					{
 						return null;
